backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
google-cloud-vision==3.4.5
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0

